Updating Parser and TutorialManager
Updated the XmlParser to be able to locate the next scene based on the current scene.
Updated tutorial manager to account for the changes to the XmlParser
Added Coroutine in TutorialManager for showing the proper content within waits in between to allow for reading time. After all text has been shown, the user is sent to the next scene until no more scenes are available.
Updated documentation in TutorialManager
Fixed spelling error in Sepsis.xml